Create README.md
Instructions
You have been selected to create a program to allow a student to store their assignment grades and check information about their current grade in the course. The specific details for the program are listed below:

The student grades program should allow the user to perform the following actions in a menu-driven format:

Add a new stored grade to the program
Display all currently stored grades
Calculate and display their current class average and letter grade
Display how many As they have scored so far
Save their progress and quit the program
Be sure to check all sections for the rubric for full points on the assignment.

Things to consider
The program should:

continually run until the user selects the Save and Quit menu option
automatically load stored grades (if any exist) from previous sessions on run
The program should NOT:

allow unexpected inputs from the user
crash if a save file does not exist or the user enters an unexpected value
Section Details
(1) Adding a new grade
Be sure the grade entered is a valid grade (0-100). You can assume the user will not try to enter a decimal number for a grade.

All grades should be stored as integers in a list.

(2) List all currently stored grades
You must use a For Loop to print each grade instead of just printing the list.

What should display if there are no grades stored?

(3) Check your course average and letter grade
What should display if there are no grades stored?

You must define and use a custom function that returns the user's average number grade.

You must define and use a custom function that returns the letter grade.

(4) Check how many As you have so far
You must use an accumulator to accomplish this.

You may also define a custom function to accomplish this task.

(5) Save and Quit
If the user chooses to quit the program, you must save their current list of grades to a file.

When running the program, your code should check for a save file and load grades from there if it exists.